refactor: extract error classes to errors.ts and use MissingSecretError in env provider
- Move MissingSecretError and SecretsProviderError to dedicated
  errors.ts to avoid circular dependency (env.ts <-> resolve.ts).
- Env provider now throws MissingSecretError instead of plain Error,
  consistent with other providers and enabling proper error handling
  in readConfigFileSnapshot().
